platform/pyproject.toml
[tool.poetry]
name = "reworkd_platform"
version = "0.1.0"
description = ""
authors = [
    "awtkns",
    "asim-shrestha"
]

maintainers = [
    "reworkd"
]

readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.98.0"
uvicorn = { version = "^0.22.0", extras = ["standard"] }
pydantic = { version = "<2.0", extras = ["dotenv"] }
ujson = "^5.8.0"
sqlalchemy = { version = "^2.0.21", extras = ["mypy", "asyncio"] }
aiomysql = "^0.1.1"
mysqlclient = "^2.2.0"
sentry-sdk = "^1.31.0"
loguru = "^0.7.2"
aiokafka = "^0.8.1"
requests = "^2.31.0"
langchain = "^0.0.295"
openai = "^0.28.0"
wikipedia = "^1.4.0"
replicate = "^0.8.4"
lanarky = "^0.7.15"
tiktoken = "^0.5.1"
grpcio = "^1.58.0"
pinecone-client = "^2.2.4"
python-multipart = "^0.0.6"
stripe = "^5.5.0"
cryptography = "^41.0.4"
httpx = "^0.25.0"
google-cloud-storage = "^2.13.0"
google-auth = "^2.23.0"


[tool.poetry.dev-dependencies]
autopep8 = "^2.0.4"
pytest = "^7.4.2"
flake8 = "~6.0.0"
mypy = "^1.5.1"
isort = "^5.12.0"
pre-commit = "^3.4.0"
wemake-python-styleguide = "^0.18.0"
black = "^23.9.1"
autoflake = "^2.2.1"
pytest-cov = "^4.1.0"
anyio = "^3.7.1"
pytest-env = "^0.8.2"

[tool.poetry.group.dev.dependencies]
dotmap = "^1.3.30"
pytest-mock = "^3.10.0"
pytest-asyncio = "^0.21.0"
mypy = "^1.4.1"
types-requests = "^2.31.0.1"
types-pytz = "^2023.3.0.0"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["reworkd_platform"]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true
exclude = "tests"

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
    "ignore::ImportWarning",
]
env = [
    "REWORKD_PLATFORM_DB_BASE=reworkd_platform_test",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
